Use awk to print line numbers in a different color
sed is not used anymore to filter the lines matched by comby. Instead, an AWK
program is generated for the same functionality, with an extra to use a
different color for line numbers.

The variable COMBY_HIGHLIGHT can be used to change the program to highlight the
source code. By default it is `pygmentize`, but it can be set to something else
like `cat` or `chroma -s tango`.

Attributes are reset before printing the line number because Chroma does left
some attributes before EOL.